Fix nits; Add test cases for writing empty data
